refactor: swagger의 custom response 폴더 구조 변경 feat: 할일 그룹 조회시 그룹의 할일 5개가 같이 조회 되도록 수정
