+++ nestjs/src/modules/todoGroup/todoGroup.module.ts
@@ -4,11 +4,13 @@ import {TodoGroupService} from './todoGroup.service';
 import {TodoModule} from "./todo/todo.module";
 import {TypeOrmModule} from "@nestjs/typeorm";
 import {TodoGroupRepository} from "./todoGroup.repository";
+import {TodoRepository} from "./todo/todo.repository";
 
 @Global()
 @Module({
     imports: [
         TypeOrmModule.forFeature([
+            TodoRepository,
             TodoGroupRepository,
         ]),
         TodoModule
@@ -17,6 +19,7 @@ import {TodoGroupRepository} from "./todoGroup.repository";
     providers: [TodoGroupService],
     exports: [
         TypeOrmModule.forFeature([
+            TodoRepository,
             TodoGroupRepository,
         ]),
         TodoGroupService,
